[9] - Creacion de el componente gif
